backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.31
asyncpg>=0.29.0
alembic>=1.13.1
WeasyPrint>=61.0
Jinja2>=3.1.3
minio==7.2.4
pydantic-settings>=2.3.4
redis>=5.0.7
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0.post1
openpyxl>=3.1.0
pdfplumber
qrcode[pil]
openai>=1.1.0
PyMuPDF>=1.24.0
